* src/lib/evil_dirent.c: 	* src/lib/evil_fnmatch.c: 	* src/lib/evil_link_xp.cpp: 	* src/lib/evil_stdio.c: 	* src/lib/evil_stdlib.c: 	* src/lib/evil_string.c: 	* src/lib/evil_unistd.c: 	* src/lib/evil_util.c: 	warnings fixes
SVN revision: 38552

+++ legacy/evil/src/lib/evil_fnmatch.c
@@ -205,7 +205,7 @@ fnmatch(const char *pattern,
 	    }
 	  else
 	    {
-	      const int	m = fnmatch_match_token(pattern + pos, *c,
+	      const size_t m = fnmatch_match_token(pattern + pos, *c,
 						leading, flags);
 
 	      if (m == FNM_SYNTAXERR)
@@ -225,7 +225,7 @@ fnmatch(const char *pattern,
 
   r = fnmatch_check_finals(pattern, states);
   fnmatch_list_of_states_free(states < new_states ? states : new_states, 2);
-  return r;
+  return (int)r;
 }
 
 #endif /* _MSC_VER || __MINGW32__ || __MINGW32CE__ */
